+++ identifykey/src/main/java/keyconstraint/identifykey/audio/Audio.java
@@ -1,9 +1,8 @@
 package keyconstraint.identifykey.audio;
 
-/** Signed 16-bit Linear PCM audio samples. */
 public interface Audio {
 
-    short[] getSamples();
+    double[] getSamples();
     int getSampleRateInHz();
     int getChannels();
     int getBitsPerSample();
